Use byte_unit::Byte in CryConfig

+++ crates/check/src/console.rs
@@ -1,4 +1,5 @@
 use anyhow::Result;
+use byte_unit::Byte;
 use std::path::Path;
 
 use cryfs_filesystem::config::Console;
@@ -40,7 +41,7 @@ impl Console for RecoverConsole {
         todo!()
     }
 
-    fn ask_blocksize_bytes_for_new_filesystem(&self) -> Result<u64> {
+    fn ask_blocksize_bytes_for_new_filesystem(&self) -> Result<Byte> {
         todo!()
     }
